+++ src/Configuration/ClientOptions.php
@@ -14,14 +14,14 @@ namespace JobRouter\AddOn\RestClient\Configuration;
 
 /**
  * Value object with available client options which can be defined
- * @see \Buzz\Client\AbstractClient->configureOptions()
+ * @see https://docs.guzzlephp.org/en/stable/request-options.html
  */
 final class ClientOptions
 {
     public function __construct(
         private readonly bool $allowRedirects = false,
         private readonly int $maxRedirects = 5,
-        private readonly int $timeout = 0,
+        private readonly int|float $timeout = 0,
         private readonly bool $verify = true,
         private readonly ?string $proxy = null,
     ) {}
@@ -29,21 +29,23 @@ final class ClientOptions
     /**
      * @internal
      * @return array{
-     *     allow_redirects: bool,
-     *     max_redirects: int,
-     *     timeout: int,
+     *     allow_redirects?: bool|array{max: int},
+     *     timeout: int|float,
      *     verify: bool,
      *     proxy: string|null
      * }
      */
     public function toArray(): array
     {
+        $allowRedirects = $this->allowRedirects ? [
+            'max' => $this->maxRedirects,
+        ] : false;
+
         /**
-         * @phpstan-ignore-next-line
+         * @phpstan-ignore-next-line Use value object over return of values
          */
         return [
-            'allow_redirects' => $this->allowRedirects,
-            'max_redirects' => $this->maxRedirects,
+            'allow_redirects' => $allowRedirects,
             'timeout' => $this->timeout,
             'verify' => $this->verify,
             'proxy' => $this->proxy,