+++ CHANGELOG.md
@@ -7,6 +7,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Changed
+- Switch to Guzzle
+
 ## [3.0.0] - 2024-02-21
 
 ### Added